Add basic logic to Cancel and Register Buttons in RegisterActivity
If hit cancel, pop notic about cancellation and return to the LoginActivity. If hit registration, if all data meet requirement, pop notic of registration successful, upload data to server (not ready yet) and direct to LoginActivity as well. If not meet all requirement, respond accordingly (coding are not ready).
For successful Registration, we can consider to direct user to MainActivity rater than LoginActivity.